Refactor common task querying into a separate func
Address some of Cale's PR comments

Combine query functions into one. No debug logging

Pull int variables out into static class variables

Mixin needs to call super __init__ too

Remove multi-course copy from upgrade reminder

Address Cale's round 2 comments
