+++ openedx/core/djangoapps/schedules/management/commands/tests/test_send_recurring_nudge.py
@@ -14,7 +14,7 @@ from mock import Mock, patch
 from opaque_keys.edx.keys import CourseKey
 from opaque_keys.edx.locator import CourseLocator
 
-from openedx.core.djangoapps.schedules import tasks
+from openedx.core.djangoapps.schedules import resolvers, tasks
 from openedx.core.djangoapps.schedules.management.commands import send_recurring_nudge as nudge
 from openedx.core.djangoapps.schedules.tests.factories import ScheduleConfigFactory, ScheduleFactory
 from openedx.core.djangoapps.site_configuration.tests.factories import SiteConfigurationFactory, SiteFactory
@@ -40,7 +40,7 @@ class TestSendRecurringNudge(CacheIsolationTestCase):
         self.site_config = SiteConfigurationFactory.create(site=site)
         ScheduleConfigFactory.create(site=self.site_config.site)
 
-    @patch.object(nudge, 'ScheduleStartResolver')
+    @patch.object(nudge.Command, 'resolver_class')
     def test_handle(self, mock_resolver):
         test_time = datetime.datetime(2017, 8, 1, tzinfo=pytz.UTC)
         nudge.Command().handle(date='2017-08-01', site_domain_name=self.site_config.site.domain)
@@ -50,7 +50,7 @@ class TestSendRecurringNudge(CacheIsolationTestCase):
             mock_resolver().send.assert_any_call(day, None)
 
     @patch.object(tasks, 'ace')
-    @patch.object(nudge, 'recurring_nudge_schedule_bin')
+    @patch.object(resolvers.ScheduleStartResolver, 'async_send_task')
     def test_resolver_send(self, mock_schedule_bin, mock_ace):
         current_time = datetime.datetime(2017, 8, 1, tzinfo=pytz.UTC)
         nudge.ScheduleStartResolver(self.site_config.site, current_time).send(-3)
@@ -80,8 +80,9 @@ class TestSendRecurringNudge(CacheIsolationTestCase):
 
         test_time = datetime.datetime(2017, 8, 3, 18, tzinfo=pytz.UTC)
         test_time_str = serialize(test_time)
-        with self.assertNumQueries(25):
-            for b in range(tasks.RECURRING_NUDGE_NUM_BINS):
+        for b in range(tasks.RECURRING_NUDGE_NUM_BINS):
+            # waffle flag takes an extra query before it is cached
+            with self.assertNumQueries(2 if b == 0 else 1):
                 tasks.recurring_nudge_schedule_bin(
                     self.site_config.site.id, target_day_str=test_time_str, day_offset=-3, bin_num=b,
                     org_list=[schedules[0].enrollment.course.org],
@@ -100,8 +101,9 @@ class TestSendRecurringNudge(CacheIsolationTestCase):
 
         test_time = datetime.datetime(2017, 8, 3, 20, tzinfo=pytz.UTC)
         test_time_str = serialize(test_time)
-        with self.assertNumQueries(25):
-            for b in range(tasks.RECURRING_NUDGE_NUM_BINS):
+        for b in range(tasks.RECURRING_NUDGE_NUM_BINS):
+            # waffle flag takes an extra query before it is cached
+            with self.assertNumQueries(2 if b == 0 else 1):
                 tasks.recurring_nudge_schedule_bin(
                     self.site_config.site.id, target_day_str=test_time_str, day_offset=-3, bin_num=b,
                     org_list=[schedule.enrollment.course.org],
@@ -124,7 +126,7 @@ class TestSendRecurringNudge(CacheIsolationTestCase):
         self.assertFalse(mock_ace.send.called)
 
     @patch.object(tasks, 'ace')
-    @patch.object(nudge, 'recurring_nudge_schedule_bin')
+    @patch.object(resolvers.ScheduleStartResolver, 'async_send_task')
     def test_enqueue_disabled(self, mock_schedule_bin, mock_ace):
         ScheduleConfigFactory.create(site=self.site_config.site, enqueue_recurring_nudge=False)
