+++ openedx/core/djangoapps/schedules/templates/schedules/edx_ace/upgradereminder/email/body.html
@@ -2,22 +2,12 @@
 {% load i18n %}
 
 {% block preview_text %}
-    {% if course_ids|length > 1 %}
-        {% blocktrans trimmed %}
-            We hope you are enjoying learning with us so far in {{ course_name }}, and other courses on
-            {{ platform_name }}! A verified certificate will allow you to highlight your new knowledge and
-            skills. It's official, and easily shareable.
+    {% blocktrans trimmed %}
+        We hope you are enjoying learning with us so far in {{ course_name }}! A verified certificate
+        will allow you to highlight your new knowledge and skills. It's official, and easily shareable.
 
-            Upgrade by {{ user_schedule_upgrade_deadline_time }}.
-        {% endblocktrans %}
-    {% else %}
-        {% blocktrans trimmed %}
-            We hope you are enjoying learning with us so far in {{ course_name }}! A verified certificate
-            will allow you to highlight your new knowledge and skills. It's official, and easily shareable.
-
-            Upgrade by {{ user_schedule_upgrade_deadline_time }}.
-        {% endblocktrans %}
-    {% endif %}
+        Upgrade by {{ user_schedule_upgrade_deadline_time }}.
+    {% endblocktrans %}
 {% endblock %}
 
 {% block content %}
@@ -26,33 +16,18 @@
         <td>
             <h1>{% trans "Upgrade now" %}</h1>
 
-            {% if course_ids|length > 1 %}
-                <p>
-                    {% blocktrans trimmed %}
-                        We hope you are enjoying learning with us so far in <strong>{{ course_name }}</strong>, and
-                        other courses on {{ platform_name }}! A verified certificate will allow you to highlight your
-                        new knowledge and skills. It's official, and easily shareable.
-                    {% endblocktrans %}
-                </p>
-                <p>
-                    {% blocktrans trimmed %}
-                        Upgrade by <strong>{{ user_schedule_upgrade_deadline_time }}</strong>.
-                    {% endblocktrans %}
-                </p>
-            {% else %}
-                <p>
-                    {% blocktrans trimmed %}
-                        We hope you are enjoying learning with us so far in <strong>{{ course_name }}</strong>! A
-                        verified certificate will allow you to highlight your new knowledge and skills. It's official,
-                        and easily shareable.
-                    {% endblocktrans %}
-                </p>
-                <p>
-                    {% blocktrans trimmed %}
-                        Upgrade by <strong>{{ user_schedule_upgrade_deadline_time }}</strong>.
-                    {% endblocktrans %}
-                </p>
-            {% endif %}
+            <p>
+                {% blocktrans trimmed %}
+                    We hope you are enjoying learning with us so far in <strong>{{ course_name }}</strong>! A
+                    verified certificate will allow you to highlight your new knowledge and skills. It's official,
+                    and easily shareable.
+                {% endblocktrans %}
+            </p>
+            <p>
+                {% blocktrans trimmed %}
+                    Upgrade by <strong>{{ user_schedule_upgrade_deadline_time }}</strong>.
+                {% endblocktrans %}
+            </p>
 
             <p>
                 <!-- email client support for style sheets is pretty spotty, so we have to inline all of these styles -->
