+++ openedx/core/djangoapps/schedules/templates/schedules/edx_ace/upgradereminder/email/body.txt
@@ -4,17 +4,6 @@
 Dear {{ user_personal_address }},
 {% endblocktrans %}
 
-{% if course_ids|length > 1 %}
-{% blocktrans trimmed %}
-    We hope you are enjoying learning with us so far in {{ course_name }}, and other courses on
-    {{ platform_name }}! A verified certificate will allow you to highlight your new knowledge and
-    skills. It's official, and easily shareable.
-
-    Upgrade by {{ user_schedule_upgrade_deadline_time }}.
-{% endblocktrans %}
-
-{% trans "Upgrade now at" %} <{{ dashboard_url }}>
-{% else %}
 {% blocktrans trimmed %}
     We hope you are enjoying learning with us so far in {{ course_name }}! A verified certificate
     will allow you to highlight your new knowledge and skills. It's official, and easily shareable.
@@ -23,4 +12,3 @@ Dear {{ user_personal_address }},
 {% endblocktrans %}
 
 {% trans "Upgrade now at" %} <{{ course_url }}>
-{% endif %}
