+++ openedx/core/djangoapps/schedules/templates/schedules/edx_ace/upgradereminder/email/from_name.txt
@@ -1,5 +1,1 @@
-{% if course_ids|length > 1 %}
-{{ platform_name }}
-{% else %}
 {{ course_name }}
-{% endif %}
